refactor: extract bot startup extensions
